+++ Sources/Geometria/Generalized/Protocols/Line/LineFloatingPoint.swift
@@ -1,6 +1,12 @@
 /// Protocol for objects that form geometric lines with two floating-point
 /// vectors representing two points on the line.
 public protocol LineFloatingPoint: LineDivisible, PointProjectableType, SignedDistanceMeasurableType where Vector: VectorFloatingPoint {
+    /// Gets the normalized slope of this line, i.e. the direction from `a -> b`.
+    var normalizedLineSlope: Vector { get }
+
+    /// Returns a generalized ``IntervalLine`` representation of this line.
+    var asIntervalLine: IntervalLine<Vector> { get }
+
     /// Performs a vector projection of a given vector with respect to this line.
     /// The resulting vector lies within the infinite line formed by extending
     /// `a <-> b`.
@@ -43,6 +49,26 @@ public protocol LineFloatingPoint: LineDivisible, PointProjectableType, SignedDi
     /// Returns a projected normalized magnitude that is guaranteed to be
     /// contained in this line.
     func clampProjectedNormalizedMagnitude(_ scalar: Magnitude) -> Magnitude
+
+    /// Returns an interval line representation of this line that is clamped such
+    /// that its end points are limited to the given minimal and maximal normalized
+    /// magnitudes relative to this line.
+    ///
+    /// Providing a minimal magnitude of `-Magnitude.infinity` or maximal
+    /// magnitude of `Magnitude.infinity` indicates that those ends of the line
+    /// should not be capped at all.
+    ///
+    /// Clamping the line with both magnitudes as their respective identity signed
+    /// infinities results in the same output as ``asIntervalLine``.
+    ///
+    /// - note: If `minimumNormalizedMagnitude <= maximumNormalizedMagnitude`,
+    /// the resulting interval line will be considered [degenerate].
+    ///
+    /// [degenerate]: https://en.wikipedia.org/wiki/Degeneracy_(mathematics)
+    func clampedAsIntervalLine(
+        minimumNormalizedMagnitude minimum: Magnitude,
+        maximumNormalizedMagnitude maximum: Magnitude
+    ) -> IntervalLine<Vector>
     
     /// Returns the squared distance between this line and a given vector.
     ///
@@ -58,6 +84,11 @@ public protocol LineFloatingPoint: LineDivisible, PointProjectableType, SignedDi
 }
 
 public extension LineFloatingPoint {
+    @inlinable
+    var normalizedLineSlope: Vector {
+        (b - a).normalized()
+    }
+
     /// Returns the closest point on this line to a given point.
     ///
     /// The point is limited to the line's bounds using
@@ -82,12 +113,41 @@ public extension LineFloatingPoint {
     
     @inlinable
     func projectedMagnitude(_ scalar: Magnitude) -> Vector {
-        a.addingProduct(lineSlope.normalized(), scalar)
+        a.addingProduct(normalizedLineSlope, scalar)
     }
     
     @inlinable
     func projectedNormalizedMagnitude(_ scalar: Magnitude) -> Vector {
         a.addingProduct(lineSlope, scalar)
+    }
+
+    @inlinable
+    func clampedAsIntervalLine(
+        minimumNormalizedMagnitude minimum: Magnitude,
+        maximumNormalizedMagnitude maximum: Magnitude
+    ) -> IntervalLine<Vector> {
+        
+        switch (minimum.isFinite, maximum.isFinite) {
+        case (false, false):
+            return asIntervalLine
+
+        case (true, true), (true, false), (false, true):
+            let clampedStart = clampProjectedNormalizedMagnitude(minimum)
+            let clampedEnd = clampProjectedNormalizedMagnitude(maximum)
+
+            if category == .lineSegment && clampedStart == .zero && clampedEnd == 1 {
+                return asIntervalLine
+            }
+
+            let slopeLength = lineSlope.length
+            
+            return IntervalLine(
+                pointOnLine: a,
+                direction: normalizedLineSlope,
+                minimumMagnitude: clampedStart * slopeLength,
+                maximumMagnitude: clampedEnd * slopeLength
+            )
+        }
     }
     
     @inlinable