+++ Tests/GeometriaTests/Generalized/IntervalLineTests.swift
@@ -3,64 +3,64 @@ import XCTest
 @testable import Geometria
 
 class IntervalLineTests: XCTestCase {
-    typealias Line = IntervalLine2D
-    typealias Line3 = IntervalLine3D
+    typealias IntervalLine = IntervalLine2D
+    typealias IntervalLine3 = IntervalLine3D
     
     func testCodable() throws {
-        let sut = Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5))
+        let sut = IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5))
         let encoder = JSONEncoder()
         let decoder = JSONDecoder()
         
         let data = try encoder.encode(sut)
-        let result = try decoder.decode(Line.self, from: data)
+        let result = try decoder.decode(IntervalLine.self, from: data)
         
         XCTAssertEqual(sut, result)
     }
     
     func testEquals() {
-        XCTAssertEqual(Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)),
-                       Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
+        XCTAssertEqual(IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)),
+                       IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
     }
     
     func testUnequals() {
-        XCTAssertNotEqual(Line(start: .init(x: 999, y: 2), end: .init(x: 3, y: 5)),
-                          Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
+        XCTAssertNotEqual(IntervalLine(start: .init(x: 999, y: 2), end: .init(x: 3, y: 5)),
+                          IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
         
-        XCTAssertNotEqual(Line(start: .init(x: 1, y: 999), end: .init(x: 3, y: 5)),
-                          Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
+        XCTAssertNotEqual(IntervalLine(start: .init(x: 1, y: 999), end: .init(x: 3, y: 5)),
+                          IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
         
-        XCTAssertNotEqual(Line(start: .init(x: 1, y: 2), end: .init(x: 999, y: 5)),
-                          Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
+        XCTAssertNotEqual(IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 999, y: 5)),
+                          IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
         
-        XCTAssertNotEqual(Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 999)),
-                          Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
+        XCTAssertNotEqual(IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 999)),
+                          IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)))
     }
     
     func testHashable() {
-        XCTAssertEqual(Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue,
-                       Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
+        XCTAssertEqual(IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue,
+                       IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
         
-        XCTAssertNotEqual(Line(start: .init(x: 999, y: 2), end: .init(x: 3, y: 5)).hashValue,
-                          Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
+        XCTAssertNotEqual(IntervalLine(start: .init(x: 999, y: 2), end: .init(x: 3, y: 5)).hashValue,
+                          IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
         
-        XCTAssertNotEqual(Line(start: .init(x: 1, y: 999), end: .init(x: 3, y: 5)).hashValue,
-                          Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
+        XCTAssertNotEqual(IntervalLine(start: .init(x: 1, y: 999), end: .init(x: 3, y: 5)).hashValue,
+                          IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
         
-        XCTAssertNotEqual(Line(start: .init(x: 1, y: 2), end: .init(x: 999, y: 5)).hashValue,
-                          Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
+        XCTAssertNotEqual(IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 999, y: 5)).hashValue,
+                          IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
         
-        XCTAssertNotEqual(Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 999)).hashValue,
-                          Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
+        XCTAssertNotEqual(IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 999)).hashValue,
+                          IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5)).hashValue)
     }
 
     func testCategory() {
-        let sut = Line(start: .zero, end: .one)
+        let sut = IntervalLine(start: .zero, end: .one)
 
         XCTAssertEqual(sut.category, .lineSegment)
     }
     
     func testAsLine() {
-        let sut = Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5))
+        let sut = IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5))
         
         let result = sut.asLine
         
@@ -73,19 +73,19 @@ class IntervalLineTests: XCTestCase {
 
 extension IntervalLineTests {
     func testA() {
-        let sut = Line(start: .zero, end: .one)
+        let sut = IntervalLine(start: .zero, end: .one)
         
         XCTAssertEqual(sut.a, .zero)
     }
     
     func testB() {
-        let sut = Line(start: .zero, end: .one)
+        let sut = IntervalLine(start: .zero, end: .one)
         
         XCTAssertEqual(sut.b, .one)
     }
 
     func testA_zeroMaximumMagnitude() {
-        let sut = Line(
+        let sut = IntervalLine(
             pointOnLine: .zero,
             direction: .one,
             minimumMagnitude: -.infinity,
@@ -96,7 +96,7 @@ extension IntervalLineTests {
     }
     
     func testB_zeroMaximumMagnitude() {
-        let sut = Line(
+        let sut = IntervalLine(
             pointOnLine: .zero,
             direction: .one,
             minimumMagnitude: -.infinity,
@@ -107,7 +107,7 @@ extension IntervalLineTests {
     }
 
     func testAB_doNotMatchIfMaximumMagnitudeIsZero() {
-        let sut = Line(
+        let sut = IntervalLine(
             pointOnLine: .zero,
             direction: .one,
             minimumMagnitude: -.infinity,
@@ -118,7 +118,7 @@ extension IntervalLineTests {
     }
 
     func testA_oneMinimumMagnitude() {
-        let sut = Line(
+        let sut = IntervalLine(
             pointOnLine: .zero,
             direction: .one,
             minimumMagnitude: 1,
@@ -129,7 +129,7 @@ extension IntervalLineTests {
     }
     
     func testB_oneMinimumMagnitude() {
-        let sut = Line(
+        let sut = IntervalLine(
             pointOnLine: .zero,
             direction: .one,
             minimumMagnitude: 1,
@@ -140,7 +140,7 @@ extension IntervalLineTests {
     }
 
     func testAB_doNotMatchIfMinimumMagnitudeIsZero() {
-        let sut = Line(
+        let sut = IntervalLine(
             pointOnLine: .zero,
             direction: .one,
             minimumMagnitude: 1,
@@ -155,7 +155,7 @@ extension IntervalLineTests {
 
 extension IntervalLineTests {
     func testOffsetBy() {
-        let sut = Line(x1: 1, y1: 2, x2: 3, y2: 5)
+        let sut = IntervalLine(x1: 1, y1: 2, x2: 3, y2: 5)
         
         let result = sut.offsetBy(.init(x: 7, y: 11))
         
@@ -169,19 +169,19 @@ extension IntervalLineTests {
 
 extension IntervalLineTests {
     func testLengthSquared() {
-        let sut = Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5))
+        let sut = IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5))
         
         XCTAssertEqual(sut.lengthSquared, 13, accuracy: 1e-14)
     }
     
     func testWithPointsScaledBy_boundedLine() {
-        let sut = Line(x1: 1, y1: 2, x2: 3, y2: 5)
+        let sut = IntervalLine(x1: 1, y1: 2, x2: 3, y2: 5)
         let factor = Vector2D(x: 7, y: 11)
         
         let result = sut.withPointsScaledBy(factor)
         
         assertEqual(
-            result.lineSlope.normalized(),
+            result.normalizedLineSlope,
             (sut.lineSlope * factor).normalized(),
             accuracy: 1e-16
         )
@@ -190,7 +190,7 @@ extension IntervalLineTests {
     }
     
     func testWithPointsScaledBy_boundedLine_negativeMinimumMagnitude() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 5)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 5)
         sut.minimumMagnitude = -5
         let factor = Vector2D(x: 1, y: 2)
         
@@ -219,7 +219,7 @@ extension IntervalLineTests {
     }
     
     func testWithPointsScaledBy_openStart() {
-        var sut = Line(x1: 1, y1: 2, x2: 3, y2: 5)
+        var sut = IntervalLine(x1: 1, y1: 2, x2: 3, y2: 5)
         sut.minimumMagnitude = -.infinity
         let factor = Vector2D(x: 7, y: 11)
         
@@ -248,7 +248,7 @@ extension IntervalLineTests {
     }
     
     func testWithPointsScaledBy_openStart_verticalLine() {
-        var sut = Line(x1: 0, y1: 0, x2: 0, y2: 5)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 0, y2: 5)
         sut.minimumMagnitude = -.infinity
         let factor = Vector2D(x: 1, y: 2)
         
@@ -277,7 +277,7 @@ extension IntervalLineTests {
     }
     
     func testWithPointsScaledBy_openStart_slantedVerticalLine() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 5)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 5)
         sut.minimumMagnitude = -.infinity
         let factor = Vector2D(x: 1, y: 2)
         
@@ -306,14 +306,14 @@ extension IntervalLineTests {
     }
     
     func testWithPointsScaledByAroundCenter_boundedLine() {
-        let sut = Line(x1: 1, y1: 2, x2: 3, y2: 5)
+        let sut = IntervalLine(x1: 1, y1: 2, x2: 3, y2: 5)
         let factor = Vector2D(x: 7, y: 11)
         let center = Vector2D(x: 5, y: 2)
         
         let result = sut.withPointsScaledBy(factor, around: center)
         
         assertEqual(
-            result.lineSlope.normalized(),
+            result.normalizedLineSlope,
             (sut.lineSlope * factor).normalized(),
             accuracy: 1e-16
         )
@@ -330,7 +330,7 @@ extension IntervalLineTests {
     }
     
     func testWithPointsScaledByAroundCenter_infiniteLine() {
-        var sut = Line(x1: 1, y1: 2, x2: 3, y2: 5)
+        var sut = IntervalLine(x1: 1, y1: 2, x2: 3, y2: 5)
         sut.minimumMagnitude = -.infinity
         sut.maximumMagnitude = .infinity
         let factor = Vector2D(x: 7, y: 11)
@@ -360,13 +360,13 @@ extension IntervalLineTests {
 
 extension IntervalLineTests {
     func testLength() {
-        let sut = Line(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5))
+        let sut = IntervalLine(start: .init(x: 1, y: 2), end: .init(x: 3, y: 5))
 
         XCTAssertEqual(sut.length, 3.605551275463989, accuracy: 1e-13)
     }
 
     func testContainsProjectedNormalizedMagnitude() {
-        let sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        let sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         
         XCTAssertFalse(sut.containsProjectedNormalizedMagnitude(-.infinity))
         XCTAssertFalse(sut.containsProjectedNormalizedMagnitude(-1))
@@ -380,7 +380,7 @@ extension IntervalLineTests {
     }
 
     func testContainsProjectedNormalizedMagnitude_nonZeroMinimumMagnitude() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         sut.minimumMagnitude = -10
         
         XCTAssertFalse(sut.containsProjectedNormalizedMagnitude(-.infinity))
@@ -395,7 +395,7 @@ extension IntervalLineTests {
     }
 
     func testContainsProjectedNormalizedMagnitude_openStart() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         sut.minimumMagnitude = -.infinity
         
         XCTAssertTrue(sut.containsProjectedNormalizedMagnitude(-.infinity))
@@ -410,7 +410,7 @@ extension IntervalLineTests {
     }
 
     func testContainsProjectedNormalizedMagnitude_openEnd() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         sut.maximumMagnitude = .infinity
         
         XCTAssertFalse(sut.containsProjectedNormalizedMagnitude(-.infinity))
@@ -425,7 +425,7 @@ extension IntervalLineTests {
     }
 
     func testContainsProjectedNormalizedMagnitude_openStartAndEnd() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         sut.minimumMagnitude = -.infinity
         sut.maximumMagnitude = .infinity
         
@@ -441,13 +441,13 @@ extension IntervalLineTests {
     }
     
     func testContainsProjectedNormalizedMagnitude_returnsFalseForNaN() {
-        let sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        let sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         
         XCTAssertFalse(sut.containsProjectedNormalizedMagnitude(.nan))
     }
     
     func testClampProjectedNormalizedMagnitude() {
-        let sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        let sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         
         XCTAssertEqual(sut.clampProjectedNormalizedMagnitude(-.infinity), 0)
         XCTAssertEqual(sut.clampProjectedNormalizedMagnitude(-1), 0)
@@ -461,7 +461,7 @@ extension IntervalLineTests {
     }
     
     func testClampProjectedNormalizedMagnitude_openStart() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         sut.minimumMagnitude = -.infinity
         
         XCTAssertEqual(sut.clampProjectedNormalizedMagnitude(-.infinity), -.infinity)
@@ -476,7 +476,7 @@ extension IntervalLineTests {
     }
     
     func testClampProjectedNormalizedMagnitude_openEnd() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         sut.maximumMagnitude = .infinity
         
         XCTAssertEqual(sut.clampProjectedNormalizedMagnitude(-.infinity), 0)
@@ -491,7 +491,7 @@ extension IntervalLineTests {
     }
     
     func testClampProjectedNormalizedMagnitude_openStartAndEnd() {
-        var sut = Line(x1: 0, y1: 0, x2: 1, y2: 0)
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 0)
         sut.minimumMagnitude = .infinity
         sut.maximumMagnitude = .infinity
         
@@ -505,32 +505,148 @@ extension IntervalLineTests {
         XCTAssertEqual(sut.clampProjectedNormalizedMagnitude(2), 2)
         XCTAssertEqual(sut.clampProjectedNormalizedMagnitude(.infinity), .infinity)
     }
-    
-    func testDistanceSquaredTo2D() {
-        let sut = Line(x1: 0, y1: 0, x2: 1, y2: 1)
-        let point = Vector2D(x: 0, y: 1)
-        
-        XCTAssertEqual(sut.distanceSquared(to: point), 0.5, accuracy: 1e-15)
+
+    func testClampedAsIntervalLine_lineMode_2D() {
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 2)
+        sut.minimumMagnitude = -.infinity
+        sut.maximumMagnitude = .infinity
+
+        let result = sut.clampedAsIntervalLine(
+            minimumNormalizedMagnitude: -10,
+            maximumNormalizedMagnitude: 20
+        )
+
+        assertEqual(
+            result.a,
+            Vector2D(x: -4.47213595499958, y: -8.94427190999916),
+            accuracy: 1e-12
+        )
+        assertEqual(
+            result.b,
+            Vector2D(x: 8.94427190999916, y: 17.88854381999832),
+            accuracy: 1e-12
+        )
+        XCTAssertEqual(sut.a.distance(to: result.a), 10.0, accuracy: 1e-13)
+        XCTAssertEqual(sut.a.distance(to: result.b), 20.0, accuracy: 1e-13)
+        XCTAssertEqual(result.length, 30.0, accuracy: 1e-13)
     }
-    
-    func testDistanceSquaredTo2D_pastStart() {
-        let sut = Line(x1: 0, y1: 0, x2: 1, y2: 1)
-        let point = Vector2D(x: -1, y: 0)
-        
-        XCTAssertEqual(sut.distanceSquared(to: point), 1, accuracy: 1e-15)
+
+    func testClampedAsIntervalLine_directionalRayMode_2D() {
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 2)
+        sut.minimumMagnitude = 0
+        sut.maximumMagnitude = .infinity
+
+        let result = sut.clampedAsIntervalLine(
+            minimumNormalizedMagnitude: -10,
+            maximumNormalizedMagnitude: 20
+        )
+
+        assertEqual(
+            result.a,
+            Vector2D(x: 0, y: 0),
+            accuracy: 1e-12
+        )
+        assertEqual(
+            result.b,
+            Vector2D(x: 8.94427190999916, y: 17.88854381999832),
+            accuracy: 1e-12
+        )
+        XCTAssertEqual(sut.a.distance(to: result.a), 0.0, accuracy: 1e-16)
+        XCTAssertEqual(sut.a.distance(to: result.b), 20.0, accuracy: 1e-14)
+        XCTAssertEqual(result.length, 20.0, accuracy: 1e-14)
     }
-    
-    func testDistanceSquaredTo2D_pastEnd() {
-        let sut = Line(x1: 0, y1: 0, x2: 1, y2: 1)
-        let point = Vector2D(x: 1, y: 2)
-        
-        XCTAssertEqual(sut.distanceSquared(to: point), 1, accuracy: 1e-15)
+
+    func testClampedAsIntervalLine_directionalRayMode_2D_invalidClamp() {
+        var sut = IntervalLine(x1: 0, y1: 0, x2: 1, y2: 2)
+        sut.minimumMagnitude = 0
+        sut.maximumMagnitude = .infinity
+
+        let result = sut.clampedAsIntervalLine(
+            minimumNormalizedMagnitude: -10,
+            maximumNormalizedMagnitude: -5
+        )
+
+        assertEqual(
+            result.a,
+            Vector2D(x: 0, y: 0),
+            accuracy: 1e-12
+        )
+        assertEqual(
+            result.b,
+            Vector2D(x: 0, y: 0),
+            accuracy: 1e-12
+        )
+        XCTAssertEqual(sut.a.distance(to: result.a), 0.0, accuracy: 1e-16)
+        XCTAssertEqual(sut.a.distance(to: result.b), 0.0, accuracy: 1e-16)
+        XCTAssertEqual(result.length, 0.0, accuracy: 1e-15)
     }
-    
-    func testDistanceSquaredTo3D() {
-        let sut = Line3(x1: 0, y1: 0, z1: 0, x2: 1, y2: 1, z2: 1)
-        let point = Vector3D(x: 1, y: 1, z: 0)
-        
-        XCTAssertEqual(sut.distanceSquared(to: point), 0.6666666666666667, accuracy: 1e-15)
+
+    func testClampedAsIntervalLine_lineSegmentMode_2D() {
+        let sut = IntervalLine(x1: 0, y1: 0, x2: 10, y2: 20)
+
+        let result = sut.clampedAsIntervalLine(
+            minimumNormalizedMagnitude: -1,
+            maximumNormalizedMagnitude: 0.5
+        )
+
+        assertEqual(
+            result.a,
+            Vector2D(x: 0, y: 0),
+            accuracy: 1e-12
+        )
+        assertEqual(
+            result.b,
+            Vector2D(x: 5.0, y: 10.0),
+            accuracy: 1e-12
+        )
+        XCTAssertEqual(sut.a.distance(to: result.a), 0.0, accuracy: 1e-16)
+        XCTAssertEqual(sut.a.distance(to: result.b), sut.length / 2.0, accuracy: 1e-14)
+        XCTAssertEqual(result.length, sut.length / 2.0, accuracy: 1e-16)
+    }
+
+    func testClampedAsIntervalLine_lineSegmentMode_2D_wholeSegment() {
+        let sut = IntervalLine(x1: 0, y1: 0, x2: 10, y2: 20)
+
+        let result = sut.clampedAsIntervalLine(
+            minimumNormalizedMagnitude: -1.0,
+            maximumNormalizedMagnitude: 2.0
+        )
+
+        assertEqual(
+            result.a,
+            Vector2D(x: 0, y: 0),
+            accuracy: 1e-12
+        )
+        assertEqual(
+            result.b,
+            Vector2D(x: 10.0, y: 20.0),
+            accuracy: 1e-12
+        )
+        XCTAssertEqual(sut.a.distance(to: result.a), 0.0, accuracy: 1e-16)
+        XCTAssertEqual(sut.a.distance(to: result.b), sut.length, accuracy: 1e-14)
+        XCTAssertEqual(result.length, sut.length, accuracy: 1e-16)
+    }
+
+    func testClampedAsIntervalLine_lineSegmentMode_2D_invalidClamp() {
+        let sut = IntervalLine(x1: 0, y1: 0, x2: 10, y2: 20)
+
+        let result = sut.clampedAsIntervalLine(
+            minimumNormalizedMagnitude: -1.0,
+            maximumNormalizedMagnitude: -0.5
+        )
+
+        assertEqual(
+            result.a,
+            Vector2D(x: 0, y: 0),
+            accuracy: 1e-50
+        )
+        assertEqual(
+            result.b,
+            Vector2D(x: 0, y: 0),
+            accuracy: 1e-50
+        )
+        XCTAssertEqual(sut.a.distance(to: result.a), 0.0, accuracy: 1e-50)
+        XCTAssertEqual(sut.a.distance(to: result.b), 0.0, accuracy: 1e-50)
+        XCTAssertEqual(result.length, 0.0, accuracy: 1e-50)
     }
 }