+++ Tests/GeometriaTests/Generalized/LineSegmentTests.swift
@@ -250,32 +250,4 @@ extension LineSegmentTests {
         XCTAssertEqual(sut.clampProjectedNormalizedMagnitude(2), 1)
         XCTAssertEqual(sut.clampProjectedNormalizedMagnitude(.infinity), 1)
     }
-    
-    func testDistanceSquaredTo2D() {
-        let sut = LineSegment(x1: 0, y1: 0, x2: 1, y2: 1)
-        let point = Vector2D(x: 0, y: 1)
-        
-        XCTAssertEqual(sut.distanceSquared(to: point), 0.5, accuracy: 1e-15)
-    }
-    
-    func testDistanceSquaredTo2D_pastStart() {
-        let sut = LineSegment(x1: 0, y1: 0, x2: 1, y2: 1)
-        let point = Vector2D(x: -1, y: 0)
-        
-        XCTAssertEqual(sut.distanceSquared(to: point), 1, accuracy: 1e-15)
-    }
-    
-    func testDistanceSquaredTo2D_pastEnd() {
-        let sut = LineSegment(x1: 0, y1: 0, x2: 1, y2: 1)
-        let point = Vector2D(x: 1, y: 2)
-        
-        XCTAssertEqual(sut.distanceSquared(to: point), 1, accuracy: 1e-15)
-    }
-    
-    func testDistanceSquaredTo3D() {
-        let sut = LineSegment3(x1: 0, y1: 0, z1: 0, x2: 1, y2: 1, z2: 1)
-        let point = Vector3D(x: 1, y: 1, z: 0)
-        
-        XCTAssertEqual(sut.distanceSquared(to: point), 0.6666666666666667, accuracy: 1e-15)
-    }
 }